Cut-over LightAspect to use OctreeByte
-the values are in the range [0..15] so this is a good fit with 7 bits so it should reduce memory usage and increase performance

+++ october-common/src/main/java/com/jeffdisher/october/aspects/LightAspect.java
@@ -9,8 +9,8 @@ import com.jeffdisher.october.types.Item;
  */
 public class LightAspect
 {
-	public static final short MAX_LIGHT = 15;
-	public static final short OPAQUE = MAX_LIGHT;
+	public static final byte MAX_LIGHT = 15;
+	public static final byte OPAQUE = MAX_LIGHT;
 
 	/**
 	 * Used to check the opacity of a block since light may only partially pass through it.  Note that all blocks have
@@ -19,9 +19,9 @@ public class LightAspect
 	 * @param item The block type.
 	 * @return The opacity value ([1..15]).
 	 */
-	public static short getOpacity(Item item)
+	public static byte getOpacity(Item item)
 	{
-		short opacity;
+		byte opacity;
 		if (ItemRegistry.AIR == item)
 		{
 			opacity = 1;
